requirements.txt
google-genai>=1.0.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pyyaml==6.0.2
numpy==2.2.2
httpx==0.28.1
cryptography>=42.0.0
Pillow>=10.0.0
opencv-python-headless>=4.9.0
markdown>=3.5.0
